EmotionalPredictionApp/requirements.txt
absl-py==2.0.0
aioice==0.9.0
aiortc==1.6.0
altair==5.2.0
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.1
av==11.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
dnspython==2.4.2
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.15.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.23.4
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.9.1
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
streamlit==1.29.0
streamlit-webrtc==0.47.1
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
